moves stations_screen to ui/stations
